# HdMergingSceneIndex: improve performance of _RebuildInputsPathTable()