--- a/pxr/imaging/hd/mergingSceneIndex.cpp
+++ b/pxr/imaging/hd/mergingSceneIndex.cpp
@@ -117,11 +117,11 @@ HdMergingSceneIndex::_RebuildInputsPathTable()
     for (auto const &inputEntry: _inputs) {
         _inputsPathTable[inputEntry.sceneRoot];
     }
-    for (auto& [path, entriesForPath]: _inputsPathTable) {
-        for (auto const &inputEntry: _inputs) {
-            if (path.HasPrefix(inputEntry.sceneRoot)) {
-                entriesForPath.push_back(inputEntry);
-            }
+
+    for (auto const &inputEntry: _inputs) {
+        auto range = _inputsPathTable.FindSubtreeRange(inputEntry.sceneRoot);
+        for (auto it = range.first; it != range.second; ++it) {
+            it->second.push_back(inputEntry);
         }
     }
 }