@jesschimein 2025-07-28
Filed as internal issue #USD-11281

(This is an automated message. See [here](https://openusd.org/release/contributing_to_usd.html#github-issues) for more information.)